Updated Sequence::sequence accessor; Removed set_sequence()
Combined the functionality of set_sequence and sequence() function into one.

If parameter is passed to sequence(), it will now assign parameter to sequence.
If parameter is not passed to sequence(), it will pass back the sequence as a string;

Added POD to Sequence_OOP.pm and examples to sequence_oop_example.pl